fix bug with autopair and iedit-rect
Autopair adds a post-command hook which completes the pair.  That hook runs
before our `iedit-reset-last-overlay', which means the paired delimiter
inserted will not be mirrored.

The fix is an alternative approaches, which does not rely a a post-command
hook.
fix bug with autopair and rectangle mode

Autopair adds a post-command hook which completes the pair.  That hook runs
before our `iedit-reset-last-overlay', which means the paired delimiter
inserted will not be mirrored.

The fix is an alternative approaches, which does not rely a a post-command
hook.